+++ src/components/user/find-login-data/index.tsx
@@ -3,7 +3,7 @@
 import { useRouter } from 'next/navigation';
 import { useShallow } from 'zustand/shallow';
 
-import LoginBackButton from '@/components/user/back-button';
+import BackButton from '@/components/layout/button/back-button';
 import FindUserId from '@/components/user/find-login-data/find-id';
 import FindPassword from '@/components/user/find-login-data/find-pw';
 import FindIdSuccess from '@/components/user/find-login-data/success';
@@ -24,7 +24,7 @@ function FindLoginData() {
 
   return (
     <>
-      <LoginBackButton onClick={backButtonHandler} />
+      <BackButton onClick={backButtonHandler} />
       {step === 1 && <FindUserId />}
       {step === 2 && <FindIdSuccess />}
       {step === 3 && <FindPassword />}
